agent/pyproject.toml
[tool.poetry]
name = "trip-advice-agent-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.1"
langchain-google-vertexai = "^2.0.11"
langchain-core = "^0.3.30"
google-cloud-aiplatform = {extras = ["langchain", "reasoningengine"], version = "^1.77.0"}
langchain-community = "^0.3.14"
langgraph = "^0.2.64"
langchain-google-firestore = "^0.5.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
